requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

boto3==1.37.13
botocore==1.37.13
chromadb==0.6.3
fastapi==0.115.9
langchain==0.3.23
langchain-chroma==0.2.2
langchain-community==0.3.21
langchain-core==0.3.51
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
numpy==1.26.4
openai==1.72.0
pillow==11.1.0
python-dotenv==1.0.1
semantic-router==0.1.1
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
